+++ app/[locale]/layout.tsx
@@ -12,8 +12,8 @@ config.autoAddCss = false;
 import "~/styles/globals.css";
 import { routing } from "~/i18n/routing";
 import { Locale } from "~/i18n/routing";
-import NavBar from "~/app/components/navbar";
-import Footer from "~/app/components/footer";
+import NavBar from "~/app/components/shared/navbar.shared";
+import Footer from "~/app/components/shared/footer.shared";
 import jwtUtil from "@/shared/utils/jwt.util";
 import { cookies } from "next/headers";
 import { Toaster } from "sonner";
