+++ app/[locale]/page.tsx
@@ -1,13 +1,13 @@
 import ROUTES from "~/constants/urls/urls";
 import ROUTES_API from "~/constants/urls/api.urls";
 import FeatureCard from "~/app/components/cards/feature.card";
-import InfluencerSlider from "~/app/components/home/influencerSlider";
+import InfluencerSlider from "~/app/components/home/influencerSlider.home";
 import Button from "~/app/components/buttons/button";
 import FireIcon from "~/app/components/icons/fire.icon";
 import Image from "next/image";
 import axios from "axios";
 import { getTranslations } from "next-intl/server";
-import NotificationSessionStorage from "~/app/components/notificationSessionStorage";
+import NotificationSessionStorage from "~/app/components/shared/notificationSessionStorage.shared";
 
 export async function generateMetadata() {
   const t = await getTranslations("HomeIndexPage");
