+++ app/components/comparison/modal.comparison.tsx
