+++ app/components/shared/filterRedirectHandler.shared.tsx
