+++ app/components/shared/footer.shared.tsx
