+++ src/domain/entities/message.entity.ts
