+++ src/interface-adapters/controllers/influencer.controller.ts
@@ -1,7 +1,7 @@
 import { influencerUseCases } from "@/application/use-cases/influencer.use-case";
 import { InfluencerOverviewPresenter } from "@/interface-adapters/presenters/influencer/influencer.overview.presenter";
 import { InfluencerDetailPresenter } from "@/interface-adapters/presenters/influencer/influencer.detail.presenter";
-import { Influencer, Status } from "@/domain/entities/influencer";
+import { Influencer, Status } from "@/domain/entities/influencer.entity";
 
 interface IndexProps {
   searchParams: {
