+++ src/interface-adapters/presenters/user/user.overview.presenter.ts
@@ -1,4 +1,4 @@
-import { User } from "@/domain/entities/user";
+import { User } from "@/domain/entities/user.entity";
 
 export class UserOverviewPresenter {
   static toHttp(user: User) {
